backend/requirements.txt
# Web Framework & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
uvicorn[standard]>=0.27.0

# Database
pymongo>=4.6.0
pydantic>=2.6.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.3
httpx>=0.26.0
curl_cffi>=0.14.0
lxml>=5.1.0

# AI
google-genai>=1.0.0

# Email & PDF
jinja2>=3.1.3
reportlab>=4.0.0

# Scheduling
apscheduler>=3.10.4

# Environment & Config
python-dotenv>=1.0.1
certifi>=2024.0.0

# Analytics & Charts
pandas>=2.0.0
plotly>=5.18.0

# Utilities
loguru>=0.7.2
